+++ site/source/components/Simulateur/ListeQuestions.tsx
@@ -7,6 +7,7 @@ import { styled } from 'styled-components'
 import { Body, Button, EditIcon, Li, Ul } from '@/design-system'
 import { Situation } from '@/domaine/Situation'
 import { GroupeDeQuestions } from '@/hooks/useQuestionsEditorialisees'
+import { useEngine } from '@/utils/publicodes/EngineContext'
 
 import { ExplicableRule } from '../conversation/Explicable'
 import Value from '../EngineValue/Value'
@@ -25,6 +26,7 @@ export const ListeQuestions = ({
 	retour,
 }: Props) => {
 	const { t } = useTranslation()
+	const engine = useEngine()
 
 	return (
 		<>
@@ -49,10 +51,14 @@ export const ListeQuestions = ({
 									</div>
 
 									<ValueContainer>
-										<Value
-											expression={premièreQuestion.id}
-											linkToRule={false}
-										/>
+										{groupe.réponse ? (
+											groupe.réponse(engine, t)
+										) : (
+											<Value
+												expression={premièreQuestion.id}
+												linkToRule={false}
+											/>
+										)}
 										<EditButton
 											light
 											onPress={() => onSélection(id)}
@@ -104,7 +110,7 @@ const BodyWithoutMargin = styled(Body)`
 `
 
 const ValueContainer = styled.div`
-	white-space: nowrap;
+	text-align: right;
 	display: flex;
 	align-items: center;
 	gap: 0 ${({ theme }) => theme.spacings.xxs};